+++ integration/e2e/limits_test.go
@@ -29,7 +29,7 @@ func TestLimits(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, c)
 
-	// should fail b/c the trace is too large
+	// should fail b/c the trace is too large. each batch should be ~80 bytes
 	batch := makeThriftBatchWithSpanCount(2)
 	require.Error(t, c.EmitBatch(context.Background(), batch))
 	// should fail b/c this will be too many traces