+++ modules/ingester/instance_test.go
@@ -58,13 +58,12 @@ func TestInstance(t *testing.T) {
 	assert.NoError(t, err, "unexpected error cutting block")
 	assert.NotEqual(t, blockID, uuid.Nil)
 
-	completeBlockID, err := i.CompleteBlock(blockID)
+	err = i.CompleteBlock(blockID)
 	assert.NoError(t, err, "unexpected error completing block")
-	assert.NotEqual(t, completeBlockID, uuid.Nil)
 
-	block := i.GetBlockToBeFlushed(completeBlockID)
+	block := i.GetBlockToBeFlushed(blockID)
 	require.NotNil(t, block)
-	assert.Len(t, i.completingBlocks, 0)
+	assert.Len(t, i.completingBlocks, 1)
 	assert.Len(t, i.completeBlocks, 1)
 
 	err = ingester.store.WriteBlock(context.Background(), block)
@@ -130,19 +129,15 @@ func TestInstanceFind(t *testing.T) {
 	pushAndQuery(t, i, request2)
 	assert.Len(t, i.completingBlocks, 2)
 
-	_, err = i.CompleteBlock(blockID)
+	err = i.CompleteBlock(blockID)
 	assert.NoError(t, err, "unexpected error completing block")
 
-	assert.Len(t, i.completingBlocks, 1)
+	assert.Len(t, i.completingBlocks, 2)
 
 	traceID := test.MustTraceID(request)
 	trace, err := i.FindTraceByID(traceID)
 	assert.NotNil(t, trace)
 	assert.NoError(t, err)
-
-	completeBlockID, err := i.CompleteBlock(blockID)
-	assert.EqualError(t, err, "error finding completingBlock")
-	assert.Equal(t, completeBlockID, uuid.Nil)
 }
 
 func TestInstanceDoesNotRace(t *testing.T) {
@@ -192,14 +187,12 @@ func TestInstanceDoesNotRace(t *testing.T) {
 	go concurrent(func() {
 		blockID, _ := i.CutBlockIfReady(0, 0, false)
 		if blockID != uuid.Nil {
-			completeBlockID, err := i.CompleteBlock(blockID)
+			err := i.CompleteBlock(blockID)
 			assert.NoError(t, err, "unexpected error completing block")
-			if completeBlockID != uuid.Nil {
-				block := i.GetBlockToBeFlushed(completeBlockID)
-				require.NotNil(t, block)
-				err := ingester.store.WriteBlock(context.Background(), block)
-				assert.NoError(t, err, "error writing block")
-			}
+			block := i.GetBlockToBeFlushed(blockID)
+			require.NotNil(t, block)
+			err = ingester.store.WriteBlock(context.Background(), block)
+			assert.NoError(t, err, "error writing block")
 		}
 	})
 
@@ -222,7 +215,7 @@ func TestInstanceDoesNotRace(t *testing.T) {
 
 func TestInstanceLimits(t *testing.T) {
 	limits, err := overrides.NewOverrides(overrides.Limits{
-		MaxSpansPerTrace: 10,
+		MaxBytesPerTrace: 1000,
 	})
 	assert.NoError(t, err, "unexpected error creating limits")
 	limiter := NewLimiter(limits, &ringCountMock{count: 1}, 1)
@@ -249,13 +242,13 @@ func TestInstanceLimits(t *testing.T) {
 			name: "succeeds",
 			pushes: []push{
 				{
-					req: test.MakeRequest(3, []byte{}),
+					req: test.MakeRequestWithByteLimit(300, []byte{}),
 				},
 				{
-					req: test.MakeRequest(5, []byte{}),
+					req: test.MakeRequestWithByteLimit(500, []byte{}),
 				},
 				{
-					req: test.MakeRequest(9, []byte{}),
+					req: test.MakeRequestWithByteLimit(900, []byte{}),
 				},
 			},
 		},
@@ -263,14 +256,14 @@ func TestInstanceLimits(t *testing.T) {
 			name: "one fails",
 			pushes: []push{
 				{
-					req: test.MakeRequest(3, []byte{}),
+					req: test.MakeRequestWithByteLimit(300, []byte{}),
 				},
 				{
-					req:          test.MakeRequest(15, []byte{}),
+					req:          test.MakeRequestWithByteLimit(1500, []byte{}),
 					expectsError: true,
 				},
 				{
-					req: test.MakeRequest(9, []byte{}),
+					req: test.MakeRequestWithByteLimit(900, []byte{}),
 				},
 			},
 		},
@@ -278,10 +271,10 @@ func TestInstanceLimits(t *testing.T) {
 			name: "multiple pushes same trace",
 			pushes: []push{
 				{
-					req: test.MakeRequest(5, []byte{0x01}),
+					req: test.MakeRequestWithByteLimit(500, []byte{0x01}),
 				},
 				{
-					req:          test.MakeRequest(7, []byte{0x01}),
+					req:          test.MakeRequestWithByteLimit(700, []byte{0x01}),
 					expectsError: true,
 				},
 			},
@@ -451,7 +444,7 @@ func TestInstanceCutBlockIfReady(t *testing.T) {
 			blockID, err := instance.CutBlockIfReady(tc.maxBlockLifetime, tc.maxBlockBytes, tc.immediate)
 			require.NoError(t, err)
 
-			_, err = instance.CompleteBlock(blockID)
+			err = instance.CompleteBlock(blockID)
 			if tc.expectedToCutBlock {
 				assert.NoError(t, err, "unexpected error completing block")
 			}