+++ pkg/flushqueues/exclusivequeues.go
@@ -3,56 +3,57 @@ package flushqueues
 import (
 	"sync"
 
-	"github.com/cortexproject/cortex/pkg/util"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/uber-go/atomic"
 )
 
 type ExclusiveQueues struct {
-	queues     []*util.PriorityQueue
+	queues     []*PriorityQueue
 	index      *atomic.Int32
 	activeKeys sync.Map
+	stopped    bool
 }
 
 // New creates a new set of flush queues with a prom gauge to track current depth
 func New(queues int, metric prometheus.Gauge) *ExclusiveQueues {
 	f := &ExclusiveQueues{
-		queues: make([]*util.PriorityQueue, queues),
+		queues: make([]*PriorityQueue, queues),
 		index:  atomic.NewInt32(0),
 	}
 
 	for j := 0; j < queues; j++ {
-		f.queues[j] = util.NewPriorityQueue(metric)
+		f.queues[j] = NewPriorityQueue(metric)
 	}
 
 	return f
 }
 
 // Enqueue adds the op to the next queue and prevents any other items to be added with this key
-func (f *ExclusiveQueues) Enqueue(op util.Op) {
+func (f *ExclusiveQueues) Enqueue(op Op) error {
 	_, ok := f.activeKeys.Load(op.Key())
 	if ok {
-		return
+		return nil
 	}
 
 	f.activeKeys.Store(op.Key(), struct{}{})
-	f.Requeue(op)
+	return f.Requeue(op)
 }
 
 // Dequeue removes the next op from the requested queue.  After dequeueing the calling
 //  process either needs to call ClearKey or Requeue
-func (f *ExclusiveQueues) Dequeue(q int) util.Op {
+func (f *ExclusiveQueues) Dequeue(q int) Op {
 	return f.queues[q].Dequeue()
 }
 
 // Requeue adds an op that is presumed to already be covered by activeKeys
-func (f *ExclusiveQueues) Requeue(op util.Op) {
+func (f *ExclusiveQueues) Requeue(op Op) error {
 	flushQueueIndex := int(f.index.Inc()) % len(f.queues)
-	f.queues[flushQueueIndex].Enqueue(op)
+	_, err := f.queues[flushQueueIndex].Enqueue(op)
+	return err
 }
 
 // Clear unblocks the requested op.  This should be called only after a flush has been successful
-func (f *ExclusiveQueues) Clear(op util.Op) {
+func (f *ExclusiveQueues) Clear(op Op) {
 	f.activeKeys.Delete(op.Key())
 }
 
@@ -69,7 +70,13 @@ func (f *ExclusiveQueues) IsEmpty() bool {
 
 // Stop closes all queues
 func (f *ExclusiveQueues) Stop() {
+	f.stopped = true
+
 	for _, q := range f.queues {
 		q.Close()
 	}
+}
+
+func (f *ExclusiveQueues) IsStopped() bool {
+	return f.stopped
 }