+++ pkg/util/test/req.go
@@ -1,9 +1,7 @@
 package test
 
 import (
-	"bytes"
 	"math/rand"
-	"sort"
 
 	"github.com/grafana/tempo/pkg/tempopb"
 	v1_common "github.com/grafana/tempo/pkg/tempopb/common/v1"
@@ -72,20 +70,35 @@ func MakeTraceWithSpanCount(requests int, spansEach int, traceID []byte) *tempop
 	return trace
 }
 
-func SortTrace(t *tempopb.Trace) {
-	sort.Slice(t.Batches, func(i, j int) bool {
-		return bytes.Compare(t.Batches[i].InstrumentationLibrarySpans[0].Spans[0].SpanId, t.Batches[j].InstrumentationLibrarySpans[0].Spans[0].SpanId) == 1
-	})
-
-	for _, b := range t.Batches {
-		sort.Slice(b.InstrumentationLibrarySpans, func(i, j int) bool {
-			return bytes.Compare(b.InstrumentationLibrarySpans[i].Spans[0].SpanId, b.InstrumentationLibrarySpans[j].Spans[0].SpanId) == 1
-		})
-
-		for _, ils := range b.InstrumentationLibrarySpans {
-			sort.Slice(ils.Spans, func(i, j int) bool {
-				return bytes.Compare(ils.Spans[i].SpanId, ils.Spans[j].SpanId) == 1
-			})
-		}
+// Note that this fn will generate a request with size **close to** maxBytes
+func MakeRequestWithByteLimit(maxBytes int, traceID []byte) *tempopb.PushRequest {
+	if len(traceID) == 0 {
+		traceID = make([]byte, 16)
+		rand.Read(traceID)
 	}
+
+	req := &tempopb.PushRequest{
+		Batch: &v1_trace.ResourceSpans{},
+	}
+
+	ils := &v1_trace.InstrumentationLibrarySpans{
+		InstrumentationLibrary: &v1_common.InstrumentationLibrary{
+			Name:    "super library",
+			Version: "0.0.1",
+		},
+	}
+	req.Batch.InstrumentationLibrarySpans = append(req.Batch.InstrumentationLibrarySpans, ils)
+
+	for req.Size() < maxBytes {
+		sampleSpan := v1_trace.Span{
+			Name:    "test",
+			TraceId: traceID,
+			SpanId:  make([]byte, 8),
+		}
+		rand.Read(sampleSpan.SpanId)
+
+		ils.Spans = append(ils.Spans, &sampleSpan)
+	}
+
+	return req
 }